Address some concerns and will finalize the scale function definition

+++ helix-core/src/main/java/org/apache/helix/controller/rebalancer/waged/constraints/SoftConstraintWeightModel.java
@@ -1,5 +1,24 @@
 package org.apache.helix.controller.rebalancer.waged.constraints;
 
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 import java.util.Map;
 
 import com.google.common.collect.ImmutableMap;
@@ -37,7 +56,7 @@ class SoftConstraintWeightModel {
    * @param originScoresMap The origin scores of soft constraints
    * @return The sum of double type
    */
-  double getSumOfScores(Map<SoftConstraint, Float> originScoresMap) {
+  float getSumOfScores(Map<SoftConstraint, Float> originScoresMap) {
     float sum = 0;
     for (Map.Entry<SoftConstraint, Float> softConstraintScoreEntry : originScoresMap.entrySet()) {
       float score = MIN_MAX_SCALER.scale(softConstraintScoreEntry.getValue());
